DesafioFinal/test-app-backend/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
PyJWT==2.3.0
django-cors-headers==3.11.0
psycopg2==2.9.0
tzdata==2022.1
pytest-django==4.5.2

